backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
chardet==5.2.0
python-multipart==0.0.6
websockets==12.0
google-cloud-translate==3.12.1
openai==1.3.0
httpx==0.25.2
